% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_all_values_are_in_set}
\alias{assert_all_values_are_in_set}
\title{Check values are in set}
\usage{
assert_all_values_are_in_set(
  test_values,
  acceptable_values,
  name = rlang::caller_arg(test_values)
)
}
\arguments{
\item{test_values}{values to check are equal to one of the values in set (character/numeric)}

\item{acceptable_values}{valid options for elements in test_values (character/numeric)}

\item{name}{name used to describe test values. Used in error message reporting i.e.  ('name' must be one of)}
}
\description{
Check if test_values are one of those described in 'set' vector
}
\examples{
set = c("A", "B", "C")
letters = c("A", "B", "B")
assert_all_values_are_in_set(letters, set)

}
