% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_is_whole_number}
\alias{assert_is_whole_number}
\title{Check object is a Whole Number}
\usage{
assert_is_whole_number(object, msg = "")
}
\arguments{
\item{object}{Some value you want to assert is a whole number (single scalar value)}

\item{msg}{Some message to print on failure (appended to the hard-coded message). Will automatically get wrapped in fmterror (string)}
}
\value{
invisible(TRUE) if the object passes the assertion. Throws an error if it does not.
}
\description{
Checks if object is a whole number (e.g. 1, 5, 5.0, 6.000). Vectors are flagged as NOT whole numbers (intentional behaviour).
}
\examples{
assert_is_whole_number(5)

}
\seealso{
Other customassertions: 
\code{\link{assert_non_empty_string}()},
\code{\link{assert_that}()}
}
\concept{customassertions}
