% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_names_include}
\alias{assert_names_include}
\title{assert_names_include}
\usage{
assert_names_include(object, expected_names, object_name_in_error_message = NA)
}
\arguments{
\item{object}{an object (usually vector or dataframe) that you want to assert has certain names}

\item{expected_names}{names you expect the object to have (order doesn't matter) (character vector)}

\item{object_name_in_error_message}{how to refer to the object in the error message}
}
\description{
assert_names_include
}
\examples{
assert_names_include(mtcars, expected_names = c("mpg", "cyl"))
}
