% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_program_exists_in_path}
\alias{assert_program_exists_in_path}
\title{Assert a program is in path}
\usage{
assert_program_exists_in_path(program_names)
}
\arguments{
\item{program_name}{name/s of program to search for in path (character)}
}
\description{
Check if program is available in path Should work on all operating systems
}
\examples{
\dontrun{
assert_program_exists_in_path(c("grep", "wget"))
}
}
