% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_that}
\alias{assert_that}
\title{Test Assertion with Invisible Return}
\usage{
assert_that(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{...}{see ?assertthat::assert_that}

\item{env}{see ?assertthat::assert_that}

\item{msg}{see ?assertthat::assert_that}
}
\value{
invisible (TRUE) if expression is TRUE. Will error if is FALSE
}
\description{
Wraps around asserthat::assert_that() but makes return value invisible (can be assigned but will not print when not assigned)
}
\seealso{
Other customassertions: 
\code{\link{assert_is_whole_number}()},
\code{\link{assert_non_empty_string}()}
}
\concept{customassertions}
