% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcheckingfunctions.R
\name{class_is}
\alias{class_is}
\title{class_is}
\usage{
class_is(object, tested_class)
}
\arguments{
\item{object}{object whose class you want to check (object)}

\item{tested_class}{class (string)}
}
\value{
TRUE if object class matches tested_class. FALSE if not.
}
\description{
Check if object has a particular class
}
