% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtfunctions.R
\name{fmterror}
\alias{fmterror}
\title{Colour text}
\usage{
fmterror(...)
}
\arguments{
\item{...}{(string/s) Text to colorise. Comma separated strings will be concatenated (no spaces) before colorisation.}
}
\value{
(string) Input text flanked by relevant Ansi escape codes
}
\description{
A collection of functions that take text and return that same text flanked by  characters
that will lead to its coloration/formatting when printed to terminals using message/cat. Different presets are available:
utilitybeltfmt::fmterror, utilitybeltfmt::fmtwarning, utilitybeltfmt::fmtsuccess, utilitybeltfmt::fmtbold.

If greater control is required, use the crayon package
}
\examples{
message(utilitybeltfmt::fmterror("This is a warning"))
}
