% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionfuntions.R
\name{fun_count_arguments}
\alias{fun_count_arguments}
\title{How Many Arguments?}
\usage{
fun_count_arguments(FUN)
}
\arguments{
\item{FUN}{a function whose arguments we're going to count (function)}
}
\value{
the number of arguments the function takes (numeric)
}
\description{
Check how many arguments a function takes.
}
\details{
Can be called from inside or outside the function, however if calling from inside a function, the function must be named.
}
\examples{
# Calling from outside a named function
fun = function(a, b, c) { return(a+b-c) }
fun_count_arguments(fun) ##Returns 3

# Calling from outside an anonymous
fun_count_arguments(function(a, b, c) { return(a+b+c) }) ##Returns 3

# Calling from inside a named function
my_function <- function(a, b, c, d, e, f, g) { return(fun_count_arguments(my_function)) }
my_function() # Equals 7
}
