% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_filename_has_valid_extension}
\alias{assert_filename_has_valid_extension}
\title{Assert file has the expected extension}
\usage{
assert_filename_has_valid_extension(filename, valid_extensions)
}
\arguments{
\item{filename}{filenames to assert has a valid extension (character)}

\item{valid_extensions}{all possible valid extensions (character)}
}
\description{
Take a filename / vector of filenames and assert that they all end with one of the user-supplied 'valid extensions'
}
\examples{
# Ensure filename has a "fasta" or 'fa' extension
assert_filename_has_valid_extension(filename="sequence.fasta", valid_extensions = c("fasta", "fa"))
}
