% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertionfunctions.R
\name{assert_files_exist}
\alias{assert_files_exist}
\title{Do all files exist}
\usage{
assert_files_exist(filepaths)
}
\arguments{
\item{filepaths}{filepaths (character)}
}
\description{
Do all files exist
}
\examples{
\dontrun{
assert_files_exist(c("/path/to/file1", "/path/to/file2"))
}
}
